+++ src/Example/Column/Formater/SimplePropertyTableColumnFormater.php
@@ -2,10 +2,9 @@
 
 namespace ILIAS\UI\Example\Table\Data\Column\Formater;
 
-use ILIAS\DI\Container;
-use ILIAS\UI\Component\Table\Data\Column\Formater\TableColumnFormater;
 use ILIAS\UI\Component\Table\Data\Column\TableColumn;
 use ILIAS\UI\Component\Table\Data\Data\Row\TableRowData;
+use ILIAS\UI\Implementation\Table\Data\Export\Formater\AbstractTableColumnFormater;
 use ILIAS\UI\Renderer;
 
 /**
@@ -15,20 +14,12 @@ use ILIAS\UI\Renderer;
  *
  * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
  */
-class SimplePropertyTableColumnFormater implements TableColumnFormater {
+class SimplePropertyTableColumnFormater extends AbstractTableColumnFormater {
 
 	/**
 	 * @inheritDoc
 	 */
-	public function __construct() {
-
-	}
-
-
-	/**
-	 * @inheritDoc
-	 */
-	public function formatHeader(TableColumn $column, Renderer $renderer, Container $dic): string {
+	public function formatHeader(TableColumn $column, Renderer $renderer): string {
 		return $column->getTitle();
 	}
 
@@ -36,7 +27,7 @@ class SimplePropertyTableColumnFormater implements TableColumnFormater {
 	/**
 	 * @inheritDoc
 	 */
-	public function formatRow(TableColumn $column, TableRowData $row, Renderer $renderer, Container $dic): string {
+	public function formatRow(TableColumn $column, TableRowData $row, Renderer $renderer): string {
 		$value = $row->getOriginalData()->{$column->getKey()};
 
 		return strval($value);